+++ Sources/LiveKit/Views/VideoView.swift
@@ -267,8 +267,6 @@ public class VideoView: NativeView, Loggable {
         _state.onDidMutate = { [weak self] newState, oldState in
             guard let self else { return }
 
-            log("Mutating in main thread: \(Thread.current.isMainThread)", .trace)
-
             let shouldRenderDidUpdate = newState.shouldRender != oldState.shouldRender
             let renderModeDidUpdate = newState.renderMode != oldState.renderMode
             let trackDidUpdate = !Self.track(oldState.track as? VideoTrack, isEqualWith: newState.track as? VideoTrack)
@@ -408,10 +406,6 @@ public class VideoView: NativeView, Loggable {
     @available(*, unavailable)
     required init?(coder _: NSCoder) {
         fatalError("init(coder:) has not been implemented")
-    }
-
-    deinit {
-        log(nil, .trace)
     }
 
     override public func performLayout() {
@@ -760,10 +754,10 @@ extension VideoView {
     static func createNativeRendererView(for renderMode: VideoView.RenderMode) -> NativeRendererView {
         #if os(iOS) || os(macOS)
         if case .sampleBuffer = renderMode {
-            logger.log("Using AVSampleBufferDisplayLayer for VideoView's Renderer", type: VideoView.self)
+            log("Using AVSampleBufferDisplayLayer for VideoView's Renderer")
             return SampleBufferVideoRenderer()
         } else {
-            logger.log("Using RTCMTLVideoView for VideoView's Renderer", type: VideoView.self)
+            log("Using RTCMTLVideoView for VideoView's Renderer")
             let result = LKRTCMTLVideoView()
 
             #if os(iOS)
@@ -780,7 +774,7 @@ extension VideoView {
                 #endif
                 // ensure it's capable of rendering 60fps
                 // https://developer.apple.com/documentation/metalkit/mtkview/1536027-preferredframespersecond
-                logger.log("preferredFramesPerSecond = 60", type: VideoView.self)
+                log("preferredFramesPerSecond = 60")
                 mtkView.preferredFramesPerSecond = 60
             }